long_island_history/backend/requirements.txt
# Long Island Historical Land Information System
# Production Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
geoalchemy2==0.14.3

# HTTP & API Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.17.2
webdriver-manager==4.0.1
playwright==1.41.0

# AI/LLM Integration
openai==1.10.0
anthropic==0.18.0
langchain==0.1.4
langchain-openai==0.0.3
langchain-anthropic==0.1.1

# GIS & Mapping
pyproj==3.6.1
shapely==2.0.2
geopandas==0.14.2
rasterio==1.3.9
owslib==0.29.3
geopy==2.4.1

# Image Processing
pillow==10.2.0
numpy==1.26.3

# Caching & Rate Limiting
redis==5.0.1
slowapi==0.1.9
cachetools==5.3.2

# Data Processing
pandas==2.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Background Tasks
celery==5.3.6
apscheduler==3.10.4

# Logging & Monitoring
structlog==24.1.0
sentry-sdk==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3
